Revert "[frameworks] Rename frameworks to now-frameworks (#3569)"
This reverts commit daa5cbdd4be7f7773a3ed40718e88c6d4f8efaab.
