+++ packages/frameworks/logos/metalsmith.png
