+++ packages/frameworks/logos/pelican.png
