+++ packages/frameworks/logos/vuepress.png
